pyproject.toml
[tool.poetry]
name = "autoos-omega"
version = "1.0.0"
description = "AUTOOS - Omega Edition: The Automation Operating System"
authors = ["AUTOOS Team"]
readme = "README.md"
packages = [{include = "autoos", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
openai = "^1.10.0"
anthropic = "^0.8.1"
google-generativeai = "^0.3.2"
chromadb = "^0.4.22"
sentence-transformers = "^2.3.1"
prometheus-client = "^0.19.0"
cryptography = "^42.0.0"
docker = "^7.0.0"
httpx = "^0.26.0"
tenacity = "^8.2.3"
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
hypothesis = "^6.96.1"
black = "^24.1.1"
mypy = "^1.8.0"
ruff = "^0.1.14"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
